POC-1/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.5.0
pydantic==2.6.0
httpx==0.27.0
sqlite3==3.41.2
